% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{areBracketsBalanced}
\alias{areBracketsBalanced}
\alias{areBracketsBalanced,FastWilcoxTest-method}
\title{check that brackets are used as should}
\usage{
areBracketsBalanced(xprs)
}
\arguments{
\item{xprs}{the string to check}
}
\description{
simply check if all brackets are opened and closed in the correct way.
}
