% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_RcppTestArea.R
\docType{methods}
\name{as_FastWilcoxTest}
\alias{as_FastWilcoxTest}
\alias{as_FastWilcoxTest,matrix-method}
\title{description of function as_FastWilcoxTest}
\usage{
\S4method{as_FastWilcoxTest}{matrix}(x)
}
\arguments{
\item{x}{a matrix object}
}
\description{
Creates a FastWilcoxTest from a matrix.
}
\examples{
set.seed(1)
ncol = 100
nrow=90
dat = matrix(round(rnorm(ncol*nrow,mean = 0, sd = 1)),ncol=ncol)
dat[which(dat < 1)] = 0
colnames(dat) <- paste('Sample', 1:ncol)
rownames(dat) <- paste( 'gene', 1:nrow)

x <- as_FastWilcoxTest( dat )
x
}
