% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_RcppTestArea.R
\docType{methods}
\name{as_RcppTestArea}
\alias{as_RcppTestArea}
\alias{as_RcppTestArea,RcppTestArea-method}
\title{description of function as_RcppTestArea}
\arguments{
\item{x}{a matrix object}
}
\description{
Creates a RcppTestArea from a matrix.
}
