% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcLinLang.R
\docType{methods}
\name{CalcLinLang}
\alias{CalcLinLang}
\alias{CalcLinLang,FastWilcoxTest-method}
\title{CalcLinLang test for rise above the detection limit}
\usage{
CalcLinLang(X, Grouping, nGroup, minPct = 0.1, display_progress = TRUE)
}
\arguments{
\item{X}{the sparse Matrix (row = cells, col = genes)}

\item{Grouping}{a numeric vector of group IDs}

\item{nGroup}{the number of groups}

\item{minPct}{ignore genes with less than 10 percent of the cells expressing them}

\item{display_progress}{show a progress bar (TRUE)}
}
\description{
Identify genes slowly rising above the detection limit
}
