% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ColNotZero}
\alias{ColNotZero}
\alias{ColNotZero,FastWilcoxTest-method}
\title{ColNotZero returns the amount of not zero values in each column}
\usage{
ColNotZero(data)
}
\arguments{
\item{data}{a sparse matrix}
}
\value{
a vector with nGene information
}
\description{
a c++ implementation of apply(x,2,function(d) {length(which(d!=0))} )
}
