% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{CorMatrix_N}
\alias{CorMatrix_N}
\alias{CorMatrix_N,FastWilcoxTest-method}
\title{Calculate correlation over two double vectors}
\usage{
CorMatrix_N(X, CMP)
}
\arguments{
\item{X}{the sparse matrix}

\item{CMP}{the vector to correlate every column of the matrix to}
}
\value{
a matrix of Rho and n (cells where both values where > 0)
}
\description{
simply calculate the correlation between X and Y
approximately 3x faster than an apply using the R cor function on sparse data
}
