% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{CorMatrixIDS}
\alias{CorMatrixIDS}
\alias{CorMatrixIDS,FastWilcoxTest-method}
\title{Calculate correlation over two double vectors}
\usage{
CorMatrixIDS(X, CMP, ids)
}
\arguments{
\item{X}{the sparse matrix}

\item{CMP}{the vector to correlate every column of the matrix to}

\item{ids}{the rows of the matrix to correlate to}
}
\description{
simply calculate the correlation between X and Y (slower than apply cor)
}
