% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{CorNormalMatrix}
\alias{CorNormalMatrix}
\alias{CorNormalMatrixIDS,FastWilcoxTest-method}
\title{Calculate correlation over two double vectors}
\usage{
CorNormalMatrix(X, CMP)
}
\arguments{
\item{X}{the normal matrix}

\item{CMP}{the vector to correlate every column of the matrix to}
}
\description{
simply calculate the correlation between X and Y
}
