% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{correlationCoefficient}
\alias{correlationCoefficient}
\alias{correlationCoefficient,FastWilcoxTest-method}
\title{Calculate correlation over two double vectors}
\usage{
correlationCoefficient(X, Y)
}
\arguments{
\item{X}{one numeric vector}

\item{Y}{the correlating vector}
}
\description{
simply calculate the correlation between X and Y
}
