% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox.R
\name{cppWilcoxTest}
\alias{cppWilcoxTest}
\title{Fast Wilcoxon rank-sum test for two vectors}
\usage{
cppWilcoxTest(x, y, type = 2L)
}
\arguments{
\item{x}{Numeric vector for the population of interest.}

\item{y}{Numeric vector for the background population.}

\item{type}{Integer test result type. `0` = greater, `1` = less,
`2` = two-sided, `3` = U statistic, `4` = -log10(greater),
`5` = log10(less), `6` = -log10(two-sided), `7` = signed
-log10(two-sided).}
}
\value{
Numeric vector containing rank sum and requested statistic/p-value.
}
\description{
Fast Wilcoxon rank-sum test for two vectors
}
