% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{eDist3d}
\alias{eDist3d}
\alias{eDist3d,FastWilcoxTest-method}
\title{find the euclidian order in a 3D vector}
\usage{
eDist3d(X, Y, Z, id, distance)
}
\arguments{
\item{X}{one numeric vector}

\item{Y}{the other vector}

\item{Z}{the thrid dimension}

\item{id}{which id to start from}

\item{distance}{the distance vector to fill up.}
}
\description{
calculates the (3D) euclidian distance for a set of x and y values
}
\details{
use the eucledian distance between one cell and all cells to find the order in the data
}
