% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{entropy}
\alias{entropy}
\alias{entropy,cellexalvrR-method}
\title{description of function entrop}
\usage{
entropy(X)
}
\arguments{
\item{X}{a double vector of group ids}
}
\description{
calculate the entropy of one double vector. Each number is a new group.
}
