% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{euclidian_distances3d}
\alias{euclidian_distances3d}
\alias{euclidian_distances3d,FastWilcoxTest-method}
\title{Calculate  over two double vectors}
\usage{
euclidian_distances3d(X, Y, Z, sum = FALSE)
}
\arguments{
\item{X}{one ORDERED numeric vector}

\item{Y}{the other vector}

\item{Z}{the thrid dimension}

\item{sum}{create a total sum of these values (integral) default = FALSE}
}
\description{
calculates the (3D) euclidian distance for a set of x and y values
}
\details{
Calculate the euclidian distance between consecutive points
Can also produce the integral values of the distance.
}
