% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian_order3d.R
\docType{methods}
\name{euclidian_order3d}
\alias{euclidian_order3d}
\alias{euclidian_order3d,FastWilcoxTest-method}
\title{description of function euclidian_order3d}
\usage{
euclidian_order3d(x, y, z, last = FALSE)
}
\arguments{
\item{x}{the x coordinate}

\item{y}{the y coordinate}

\item{z}{the z coordinate}

\item{last}{not re-sample the data (default = FALSE)}
}
\description{
identify a order in a linear set of 3D coordinates.
}
