% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_proximity_oob}
\alias{extract_proximity_oob}
\alias{extract_proximity_oob,FastWilcoxTest-method}
\title{extract proximity for the ranger results
copied from https://github.com/imbs-hl/ranger/issues/234}
\usage{
extract_proximity_oob(pred, prox, inbag)
}
\arguments{
\item{pred}{the predictions created from a predict(rangerRF, data, type = "terminalNodes")$predictions}

\item{prox}{an empty matrix with dim(pred) dimensions}

\item{inbag}{the inbag information from the ranger prediction run (rangerRF$inbag.counts)}
}
\value{
prox with correct values
}
\description{
calculate the proximity matrix
}
