% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_class.R
\docType{class}
\name{FastWilcoxTest-class}
\alias{FastWilcoxTest-class}
\title{FastWilcoxTest}
\description{
An S4 class to train to code R C++
The class itself is totally unused. All function are directly applied to the dgCMatrix matrix.
}
\details{
The FastWilcoxTest class

is only a wrapper around a sparse matrix. The class itself is not used apart from the test scripts.
All functionallity is exported as c++ header FastWilcoxTest.h 
as well as normal R function FastWilcosTest::<function name>().
}
\section{Slots}{

\describe{
\item{\code{dat}}{a dgCMatrix to learn how to modify that using C++}

\item{\code{usedObj}}{here a set of used and probably important objects can be saved. Be very careful using any of them!}
}}

