% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{LogNorm}
\alias{LogNorm}
\alias{LogNorm,FastWilcoxTest-method}
\title{UMI normalize a single cell expression matrix}
\usage{
LogNorm(data, scale_factor, display_progress = TRUE)
}
\arguments{
\item{scale_factor}{the total read count to reach}

\item{display_progress}{show a progress bar (TRUE)}

\item{X}{the sparse Matrix (row = genes, col = cells)}
}
\description{
Normalize the single cell expression read counts values to a scale_factor.
}
