% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{MEAN_STD}
\alias{MEAN_STD}
\alias{MEAN_STD,FastWilcoxTest-method}
\title{Calculate mean and std of >0 values in a sparse matrix}
\usage{
MEAN_STD(data)
}
\arguments{
\item{data}{a spare matrix}
}
\description{
A specific z. score method that converts the data to 10 +-1 instead of 0+-1
in order to keep the not expressed clearly separate from the real data.
}
