% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltSparseMatrix.R
\docType{methods}
\name{meltSparseMatrix}
\alias{meltSparseMatrix}
\alias{meltSparseMatrix,FastWilcoxTest-method}
\title{description of function meltSparseMatrix}
\arguments{
\item{x}{a sparse matrix}
}
\value{
a data.frame with the columns gene_id, cell_id and value
containing the row ids, the col ids and the value stored in the sparse matrix using c++
}
\description{
melt a sparse matrix into row ids, col ids and value
}
