% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{NormalizeCells}
\alias{NormalizeCells}
\alias{NormalizeCells,FastWilcoxTest-method}
\title{UMI normalize a single cell expression matrix}
\usage{
NormalizeCells(X, nUMI, display_progress = TRUE)
}
\arguments{
\item{nUMI}{aim to normalize each cell to (cells expressing less are set to 0)}

\item{display_progress}{show a progress bar (TRUE)}

\item{data}{the sparse Matrix (row = genes, col = cells)}
}
\description{
Normalize the single cell expression values to a total of nUMI reads.
}
