% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{NormalizeSamples}
\alias{NormalizeSamples}
\alias{NormalizeSamples,FastWilcoxTest-method}
\title{rescale a matrix using scaleFactor per column}
\usage{
NormalizeSamples(X, scaleFactor, display_progress = TRUE)
}
\arguments{
\item{scaleFactor}{the geometric mean of the data (use DESeq2 to get these)}

\item{display_progress}{show a progress bar (TRUE)}

\item{data}{the sparse Matrix (row = genes, col = cells)}
}
\description{
Normalize the NGS expression values by log re-scaling.
}
