% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/show.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,RcppTestArea-method}
\alias{show}
\alias{show,RcppTestArea-method}
\title{description of function print}
\arguments{
\item{x}{the RcppTestArea object}

\item{object}{the RcppTestArea object}
}
\value{
nothing

nothing
}
\description{
print the RcppTestArea

show the RcppTestArea
}
