% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_class.R
\docType{class}
\name{FastWilcoxTest}
\alias{FastWilcoxTest}
\title{FastWilcoxTest}
\description{
An S4 class to train to code R C++
}
\section{Slots}{

\describe{
\item{\code{dat}}{a dgCMatrix to lear how to modify that using C++}

\item{\code{usedObj}}{here a set of used and probably lateron important objects can be saved. Be very carful using any of them!}
}}

