% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{rgb2hex}
\alias{rgb2hex}
\alias{rgb2hex,FastWilcoxTest-method}
\title{hex color vector 4 RGB matrix}
\usage{
rgb2hex(Mat)
}
\arguments{
\item{Mat}{the matrix or RGB values (numeric)}
}
\value{
hex color
}
\description{
return a vector of hex values for a matrix of RGB values (rows)
}
