% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{rgb2hexS}
\alias{rgb2hexS}
\alias{rgb2hexS,FastWilcoxTest-method}
\title{hex color RGB values (int)}
\usage{
rgb2hexS(red, green, blue)
}
\arguments{
\item{red}{int value (0..255)}

\item{green}{int value (0..255)}

\item{blue}{int value (0..255)}
}
\value{
string vector of nrow(Mat) length
}
\description{
return a vector of hex values for a matrix of RGB values (rows)
}
