% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{rollSum}
\alias{rollSum}
\alias{rollSum,FastWilcoxTest-method}
\title{rolling sum over sparse matrix}
\usage{
rollSum(X, n)
}
\arguments{
\item{X}{the sparse matrix}

\item{n}{the size of the rolling window}
}
\description{
calculate a rolling sum of the rows
}
