% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{rollSumStart}
\alias{rollSumStart}
\alias{rollSumStart,FastWilcoxTest-method}
\title{rolling sum over sparse matrix}
\usage{
rollSumStart(X, n, S)
}
\arguments{
\item{X}{the sparse matrix}

\item{n}{the length of the rolling window}

\item{S}{the start positions of each X row}
}
\description{
calculate a rolling sum of the rows
}
\details{
Calculatethe rolling sum for a max distance from the start
The location of each row is taken from the S (Start) vector
In R the colnames need to be set to the input colnames whereas the start positions need to be set to chrXY : S[i] - (S[i]+n)
}
