% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstats.R
\docType{methods}
\name{Rstats}
\alias{Rstats}
\alias{Rstats,FastWilcoxTest-method}
\title{description of function Rstats}
\usage{
Rstats(X, interest, backgound, logFCcut = 1)
}
\arguments{
\item{X}{the sparse matrix}

\item{interest}{the intereting col IDs}

\item{backgound}{the background col IDs}

\item{logFCcut}{the logFC change cutoff default= 1.0}
}
\description{
calculate the wilcox test on a sparse matrix as the C++ function does it.
The data in the sparse matrix should be log transformed (not log10 or log2!)
}
