% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstats.R
\docType{methods}
\name{Rstats,matrix-method}
\alias{Rstats,matrix-method}
\title{description of function Rstats}
\usage{
\S4method{Rstats}{matrix}(X, interest, backgound, logFCcut = 1,
  minPct = 0.1, onlyPos = FALSE)
}
\arguments{
\item{X}{the normal matrix}

\item{interest}{the interesting col IDs}

\item{backgound}{the background col IDs}

\item{logFCcut}{the logFC change cutoff default= 1.0}

\item{minPct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function}

\item{onlyPos}{test only those genes with higher expression in the interest (default FALSE)}
}
\description{
use the R wilcox.test function instead of a c++ version.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: matrix
}}

