% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rzscore.R
\docType{methods}
\name{Rzscore}
\alias{Rzscore}
\alias{Rzscore,FastWilcoxTest-method}
\title{description of function Rzscore}
\usage{
Rzscore(obj)
}
\arguments{
\item{obj}{the FastWilcoxTest}
}
\description{
calculate the z.score as the cpp function does it
}
