% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rzscore.R
\docType{methods}
\name{RzscoreAll}
\alias{RzscoreAll}
\alias{RzscoreAll,FastWilcoxTest-method}
\title{description of function Rzscore}
\arguments{
\item{obj}{the FastWilcoxTest}
}
\description{
calculate the z.score as the cpp function does it
}
