% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,FastWilcoxTest-method}
\title{description of function show}
\arguments{
\item{object}{the FastWilcoxTest object}
}
\value{
nothing
}
\description{
show the FastWilcoxTest
}
