% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,FastWilcoxTest-method}
\title{description of function show}
\usage{
\S4method{show}{FastWilcoxTest}(object)
}
\arguments{
\item{object}{the FastWilcoxTest object}
}
\description{
show the FastWilcoxTest
}
