% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparse2SQLite_text_file}
\alias{sparse2SQLite_text_file}
\alias{sparse2SQLite_text_file,FastWilcoxTest-method}
\title{sparse2SQLite_text_file creates a simple text file from the matrix contents}
\usage{
sparse2SQLite_text_file(data, file, sep = " ")
}
\arguments{
\item{data}{a sparse matrix}
}
\value{
a vector with nGene information
}
\description{
circumvent the memory expenses during RSQlite database creation (melting the whole matrix)
}
