% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{SphericEntropy}
\alias{SphericEntropy}
\alias{SphericEntropy,cellexalvrR-method}
\title{description of function entrop}
\usage{
SphericEntropy(X1, X2, X3, gvect, radii)
}
\arguments{
\item{X1}{a double vector of dim1}

\item{X2}{a double vector of dim2}

\item{X3}{a double vector of dim3}

\item{gvect}{a double vector of the output group}

\item{radii}{a double vector of the euclidian max distances (radius) to be checked}
}
\description{
Calculate a per cell entropy for a vector of radius cut-offs.
}
