% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{toColNums}
\alias{toColNums}
\alias{toColNums,FastWilcoxTest-method}
\title{toColNums returns a vector with column IDs}
\usage{
toColNums(data)
}
\arguments{
\item{data}{a sparse matrix}
}
\value{
a vector with nGene information

a vector of col ids in the order of the @x vector
}
\description{
using c++ to get the column IDS mapping to the @x values.
}
