% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_class.R, R/data.R
\docType{data}
\name{x}
\alias{x}
\title{A simple FastWilcoxTest object containing random data to run the tests on.}
\format{
FastWilcoxTest

a sparse matrix (90 x 100) as a FastWilcoxTest object.
}
\usage{
x

x
}
\description{
The data was created following the example in as_FastWilcoxTest()
}
\keyword{data}
