% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{ZScore}
\alias{ZScore}
\alias{ZScore,FastWilcoxTest-method}
\title{Calculate z score for a sparse matrix}
\usage{
ZScore(data, display_progress = TRUE)
}
\arguments{
\item{data}{the sparse Matrix}

\item{display_progress}{show a progress bar (TRUE)}
}
\description{
A specific z. score method that converts the data to 10 +-1 instead of 0+-1
in order to keep the not expressed clearly separate from the real data.
}
