% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{ZScoreAll}
\alias{ZScoreAll}
\alias{ZScoreAll,FastWilcoxTest-method}
\title{Calculate z score for a sparse matrix}
\usage{
ZScoreAll(data, display_progress = TRUE)
}
\arguments{
\item{data}{the sparse Matrix}

\item{display_progress}{show a progress bar (TRUE)}
}
\description{
A sparse matrix z. score function returning the same as a normal one
}
