% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_RcppTestArea.R
\docType{methods}
\name{as_FastWilcoxTest}
\alias{as_FastWilcoxTest}
\alias{as_FastWilcoxTest,matrix-method}
\title{description of function as_FastWilcoxTest}
\usage{
\S4method{as_FastWilcoxTest}{matrix}(x)
}
\arguments{
\item{x}{a matrix object}
}
\description{
Creates a FastWilcoxTest from a matrix.
}
