% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_FastWilcoxTest.R
\docType{methods}
\name{as_FastWilcoxTest}
\alias{as_FastWilcoxTest}
\alias{as_FastWilcoxTest,FastWilcoxTest-method}
\title{description of function as_FastWilcoxTest}
\arguments{
\item{x}{a matrix object}
}
\description{
Creates a FastWilcoxTest from a matrix.
}
