% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppWilcoxTest}
\alias{cppWilcoxTest}
\alias{cppWilcoxTest,cppWilcoxTest-method}
\title{cppWilcoxTest runs wilcox test comparing x and y}
\usage{
cppWilcoxTest(x, y, type)
}
\arguments{
\item{x}{the interesting information}

\item{y}{the background information}

\item{type}{the test type (greater=0,less=1,twoSided=2,U=3,Abslog10greater=4,log10less=5,abslog10twoSided=6,Q=7)}
}
\value{
a double vector with 2 entries : Rank sum and p.value,
}
\description{
a simple replacement of wilcox.test returning less information but >10x faster
}
