% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/euclidian_order3d.R
\docType{methods}
\name{euclidian_order3d}
\alias{euclidian_order3d}
\alias{euclidian_order3d,FastWilcoxTest-method}
\title{find the euclidian order in a 3D vector}
\usage{
euclidian_order3d(x, y, z, last = FALSE)

euclidian_order3d(x, y, z, last = FALSE)
}
\arguments{
\item{x}{the x coordinate}

\item{y}{the y coordinate}

\item{z}{the z coordinate}

\item{last}{not re-sample the data (default = FALSE)}

\item{X}{one numeric vector}

\item{Y}{the other vector}

\item{Z}{the thrid dimension}
}
\description{
calculates the (3D) euclidian distance for a set of x and y values

identify a order in a linear set of 3D coordinates.
}
\details{
use the eucledian distance between all cells to find the order in the data
}
