% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FC}
\alias{FC}
\alias{FC,FastWilcoxTest-method}
\title{FC calculates a log fold change between the two input vectors}
\usage{
FC(A, B)
}
\arguments{
\item{A}{one numeric vector of log data}

\item{B}{the other log vector}
}
\value{
a double fold change
}
\description{
a simple replacement of wilcox.test returning less information but >10x faster
}
