% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltSparseMatrix.R
\docType{methods}
\name{meltSparseMatrix}
\alias{meltSparseMatrix}
\alias{meltSparseMatrix,dgCMatrix-method}
\title{fast melt a sparse matrix omitting all '0' entries.}
\usage{
\S4method{meltSparseMatrix}{dgCMatrix}(x)
}
\arguments{
\item{x}{a sparse matrix}
}
\description{
melt a sparse matrix into a data.frame with gene.id = row ids, cell_id = col ids and value = x@x
}
\examples{
x
melted = meltSparseMatrix(x@dat)
dim(melted)
colnames(melted)
}
