% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,FastWilcoxTest-method}
\title{description of function print}
\usage{
\S4method{print}{FastWilcoxTest}(x)
}
\arguments{
\item{x}{the FastWilcoxTest object}
}
\value{
nothing
}
\description{
print the FastWilcoxTest
}
