% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/show.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,FastWilcoxTest-method}
\alias{show}
\alias{show,FastWilcoxTest-method}
\title{description of function print}
\arguments{
\item{x}{the FastWilcoxTest object}

\item{object}{the FastWilcoxTest object}
}
\value{
nothing

nothing
}
\description{
print the FastWilcoxTest

show the FastWilcoxTest
}
