% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{rollSumStart}
\alias{rollSumStart}
\alias{rollSum}
\alias{rollSumStart,FastWilcoxTest-method}
\title{rolling sum over sparse matrix}
\usage{
rollSum(X, n)
}
\arguments{
\item{X}{the sparse matrix}

\item{n}{the length of the rolling window}

\item{S}{the start positions of each X row}
}
\description{
calculate a rolling sum of the rows
}
\details{
Calculatethe rolling sum for a max distance from the start
}
