% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstats.R
\docType{methods}
\name{Rstats}
\alias{Rstats}
\alias{Rstats,FastWilcoxTest-method}
\title{description of function Rstats}
\usage{
Rstats(X, interest, backgound, logFCcut = 1, minPct = 0.1,
  onlyPos = FALSE)
}
\arguments{
\item{X}{the sparse matrix}

\item{interest}{the intereting col IDs}

\item{backgound}{the background col IDs}

\item{logFCcut}{the logFC change cutoff default= 1.0}

\item{minPct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function}

\item{onlyPos}{test only those genes with higher expression in the interest (default FALSE)}
}
\description{
calculate the wilcox test on a sparse matrix as the C++ function does it.
The data in the sparse matrix should be log transformed (not log10 or log2!)
}
