% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rzscore.R
\docType{methods}
\name{Rzscore}
\alias{Rzscore}
\alias{Rzscore,FastWilcoxTest-method}
\title{description of function Rzscore}
\arguments{
\item{obj}{the FastWilcoxTest}
}
\description{
calculate the z.score as the cpp function does it
}
\examples{
x
zscored = Rstats(x@dat, interest=1:10, background=20:30, logFCcut= 0, minPct=0, onlyPos=FALSE )
dim(zscored)
}
