% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rzscore.R
\docType{methods}
\name{Rzscore}
\alias{Rzscore}
\alias{Rzscore,RcppTestArea-method}
\title{description of function Rzscore}
\arguments{
\item{obj}{the RcppTestArea}
}
\description{
calculate the z.score as the cpp function does it
}
