% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,BioData-method}
\title{description of function show}
\arguments{
\item{x}{the BioData object}

\item{object}{TEXT MISSING}
}
\value{
nothing
}
\description{
show the BioData
}
