% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ShuffleMatrix}
\alias{ShuffleMatrix}
\alias{ShuffleMatrix,FastWilcoxTest-method}
\title{reshuffle data based on a sparse matrix assuming max double the amount of entries not being zero}
\usage{
ShuffleMatrix(X)
}
\arguments{
\item{X}{the sparse matrix (tests are applied to columns!)}
}
\value{
a matrix with x, j and i avalues to be put into a new sparse matrix
}
\description{
replacing the synthetic1 function of RFclust.SGE package
}
