% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StatTest}
\alias{StatTest}
\alias{StatTest,FastWilcoxTest-method}
\title{StatTest runs wilcox test on the columns of the sparse matrix}
\usage{
StatTest(X, interest, background, logFCcut = 1, minPct = 0.1)
}
\arguments{
\item{X}{the sparse matrix (tests are applied to columns!)}

\item{interest}{row IDs for the group of interest}

\item{background}{row IDS for the background}

\item{logFCcut}{data is meant to be log() transformed and only columns passing a logFCcut of (default 1) are tested}

\item{minPct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.1}
}
\value{
a matrix with tested column ids, logFC and p.value
}
\description{
This test implements the Seurat FindMarkers( test.use == "wilcox" ) function
in the greatest possible way, but using Rcpp instead of R. So far I could get a ~10x speed improvement.
}
